Use kmp native coroutines and clean up visibility
This PR cleans up the visibility of various Kotlin functions to avoid
certain methods (ex the specific underlying database methods) from being
called from Android/Swift apps. It also adds the dependency and
annotations for kmp-nativecoroutines.
