+++ gradle/libs.versions.toml
@@ -3,7 +3,10 @@ agp = "8.12.0"
 kotlin = "2.2.21"
 kotlinx-serialization = "1.9.0"
 kotlinx-datetime = "0.7.1-0.6.x-compat"
+ksp = "2.3.0"
+
 kermit = "2.0.8"
+native-coroutines = "1.0.0-ALPHA-47"
 
 maven-publish = "0.34.0"
 
@@ -70,3 +73,5 @@ kotlin-multiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref 
 kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin" }
 sqldelight = { id = "app.cash.sqldelight", version.ref = "sqldelight" }
 vanniktech-maven-publish = { id = "com.vanniktech.maven.publish", version.ref = "maven-publish" }
+ksp = { id = "com.google.devtools.ksp", version.ref = "ksp" }
+native-coroutines = { id = "com.rickclephas.kmp.nativecoroutines", version.ref = "native-coroutines" }
