+++ persistence/src/commonMain/kotlin/com/quran/shared/persistence/model/DatabaseTypes.kt
@@ -2,4 +2,4 @@ package com.quran.shared.persistence.model
 
 import com.quran.shared.persistence.Page_bookmark
 
-typealias DatabasePageBookmark = Page_bookmark
+internal typealias DatabasePageBookmark = Page_bookmark
